requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.7
django-admin-interface==0.29.4
django-appconf==1.0.6
django-cachalot==2.7.0
django-celery-results==2.5.1
django-colorfield==0.11.0
django-countries==7.6.1
django-debug-toolbar==5.0.1
django-filter==24.3
django-image-cropping==1.7
django-recaptcha==4.0.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
easy-thumbnails==2.9
gunicorn==23.0.0
inflection==0.5.1
kombu==5.4.2
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
setuptools==75.1.0
six==1.17.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13